Move dartdoc_page.dart to _pub_shared.
