+++ pkg/_pub_shared/pubspec.yaml
@@ -8,6 +8,7 @@ dependencies:
   html: ^0.15.0
   json_annotation: ^4.3.0
   meta: ^1.3.0
+  sanitize_html: ^2.1.0
   api_builder:
     path: ../api_builder
